Add --client test, clean out anonymous users
Adding a test for `--client` revealed that we're not cleaning out
anonymous users. They're not really a security concern (there's nothing
else running with MySQL in its container), they're not useful either,
and can lead to confusing errors when attempting to login as the aptible
user from localhost (because they'll be matches first).