Stop using mysqld_safe
There is hardly anything in mysqld_safe that we use, but there is plenty
we *don't* use.

- First, mysqld_safe is a shell script that runs as PID 1 and
  deliberately ignores all the signals it receives. This is bad for us,
  because we want to be able to issue a `docker restart` to a MySQL
  container. If we keep using mysqld_safe, Docker's SIGTERM will be
  ignored, and we'll receive a SIGKILL.
- Second, mysqld_safe captures error logging output and routes it to a
  file. Once again, we don't want that. We'd rather have it sent to
  stdout.
- Finally, mysqld_safe essentially parses the options from our options
  file to pass them to MySQL as arguments. It's unclear why that's
  useful (we have test doverage for the options we set). It also does a
  ton of other things we don't need, like being compatible with RHEL 5,
  sh, etc., which we don't actually use.

There are two reasons left why mysqld_safe might be useful to us:o

- It restarts MySQL when it dies
- It cleans out old PID files

The latter doesn't seem to be needed (we now have tests for this), and
I'll investigate running databases to check whether the former is
useful.

All in all, mysqld_safe is arguably a tool for another era, and using it
in our Docker image makes little sense (incidentally, Docker does not
use it in their own image).