Configure innodb_log_file_size
The log file size default on MySQL is not really appropriate for any
workload, so we turn that up a bit, from 5MB to 256MB.

Note that the log file size cannot be changed after the database has
been started, so we only configure this file when running --initialize.

This is fairly costly in terms of disk space, but considering all MySQL
databases on Enclave have 10GB of disk anyway (and our plans include a
large amount of free disk space), it's probably what's best for our
customers (of course, every workload would have its own sweet spot, and
some customers are likely to need less, but this is the best we can do
as a basic heuristic).

See: https://www.percona.com/blog/2011/11/21/should-mysql-update-the-default-innodb_log_file_size/