Move logging tests outside of the container
For reasons that aren't clear, running the logging tests on Travis now
fails. The issue appears to be related to buffering in tail or sed,
since the relevant logs *are* found in e.g. in `.../general.log` and
`.../slow.log`, but not in the main MySQL logs. The issue is not
reproducible locally (on OSX or Linux).

This patch modifies our tests so these tests run outside of the
container by looking at `docker logs` instead. In practice, this makes
the tests more robust, so it's probably a net win anyway.