Skip test-log.sh on Travis
Reproducing the issue on Travis, it appears inotify does not notify tail
when tailing the logs in the container (I concluded that by running
`strace` on `tail` while MySQL was writing to `general.log`).

This used to work, and more importantly, it works fine it we push the
image we use on Travis to other locations (my local workstation and an
Enclave test instance). So, for now, there's no better option than to
skip this test on Travis.

We'll revisit in a little while to see if this was a temporary problem
with Travis. If not, I'll try to prepare a minimal repro.